refactor(tests): moved `create_test_socket` into test module
Moved `create_test_socket` into test module
