+++ src/vmm/src/devices/virtio/vhost_user.rs
@@ -462,11 +462,11 @@ impl<T: VhostUserHandleBackend> VhostUserHandleImpl<T> {
 #[cfg(test)]
 mod tests {
     #![allow(clippy::undocumented_unsafe_blocks)]
-    #![allow(clippy::cast_possible_truncation)]
 
     use utils::tempfile::TempFile;
 
     use super::*;
+    use crate::utilities::test_utils::create_tmp_socket;
     use crate::vstate::memory::{FileOffset, GuestAddress, GuestMemoryExtension};
 
     #[test]
@@ -494,40 +494,7 @@ mod tests {
 
         let max_queue_num = 69;
 
-        let tmp_dir = utils::tempdir::TempDir::new().unwrap();
-        let tmp_dir_path_str = tmp_dir.as_path().to_str().unwrap();
-        let tmp_socket_path = format!("{tmp_dir_path_str}/tmp_socket");
-
-        unsafe {
-            let socketfd = libc::socket(libc::AF_UNIX, libc::SOCK_STREAM, 0);
-            if socketfd < 0 {
-                panic!("Cannot create socket");
-            }
-            let mut socket_addr = libc::sockaddr_un {
-                sun_family: libc::AF_UNIX as u16,
-                sun_path: [0; 108],
-            };
-
-            std::ptr::copy::<i8>(
-                tmp_socket_path.as_ptr().cast(),
-                socket_addr.sun_path.as_mut_ptr(),
-                tmp_socket_path.as_bytes().len(),
-            );
-
-            let bind = libc::bind(
-                socketfd,
-                (&socket_addr as *const libc::sockaddr_un).cast(),
-                std::mem::size_of::<libc::sockaddr_un>() as u32,
-            );
-            if bind < 0 {
-                panic!("Cannot bind socket");
-            }
-
-            let listen = libc::listen(socketfd, 1);
-            if listen < 0 {
-                panic!("Cannot listen on socket");
-            }
-        }
+        let (_tmp_dir, tmp_socket_path) = create_tmp_socket();
 
         // Creation of the VhostUserHandleImpl correctly connects to the socket, sets the maximum
         // number of queues and sets itself as an owner of the session.
